Add TensorMut and TensorRef tensor contains
